+++ tgbotapi.core/src/jsMain/kotlin/dev/inmo/tgbotapi/requests/abstracts/MPPFileMultipartFileActual.kt
@@ -1,0 +1,8 @@
+package dev.inmo.tgbotapi.requests.abstracts
+
+import dev.inmo.micro_utils.common.*
+import io.ktor.utils.io.ByteReadChannel
+
+actual suspend fun MPPFile.asMultipartFile(): MultipartFile = ByteReadChannel(bytes()).asMultipartFile(
+    filename.name
+)