+++ tgbotapi.core/src/jvmMain/kotlin/dev/inmo/tgbotapi/requests/abstracts/MPPFileMultipartFileActual.kt
@@ -1,0 +1,5 @@
+package dev.inmo.tgbotapi.requests.abstracts
+
+import dev.inmo.micro_utils.common.MPPFile
+
+actual suspend fun MPPFile.asMultipartFile(): MultipartFile = toInputFile()